RAG_PDF_Chatbot_Gemini/requirements.txt
streamlit
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
google-generativeai
python-dotenv
langchain